third_party/blink/web_tests/platform/mac/external/wpt/css/css-text/white-space/seg-break-transformation-004.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] linebreak only ₩２４
  assert_equals: expected 73 but got 79
[FAIL] spaces linebreak ₩２４
  assert_equals: expected 73 but got 79
[FAIL] linebreak spaces ₩２４
  assert_equals: expected 73 but got 79
[FAIL] spaces linebreak spaces ₩２４
  assert_equals: expected 73 but got 79
[FAIL] multiple linebreaks ₩２４
  assert_equals: expected 73 but got 79
[FAIL] multiple linebreaks + spaces ₩２４
  assert_equals: expected 73 but got 79
[FAIL] linebreak only ２４₩
  assert_equals: expected 73 but got 79
[FAIL] spaces linebreak ２４₩
  assert_equals: expected 73 but got 79
[FAIL] linebreak spaces ２４₩
  assert_equals: expected 73 but got 79
[FAIL] spaces linebreak spaces ２４₩
  assert_equals: expected 73 but got 79
[FAIL] multiple linebreaks ２４₩
  assert_equals: expected 73 but got 79
[FAIL] multiple linebreaks + spaces ２４₩
  assert_equals: expected 73 but got 79
Harness: the test ran to completion.
